+++ doc.go
@@ -35,7 +35,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-8.1.3
+8.2.5
 
 Installation
 
@@ -680,6 +680,94 @@ Example code:
     }
 
 
+Controllers
+
+It's very easy to get started, the only function you need to call
+instead of `app.Get/Post/Put/Delete/Connect/Head/Patch/Options/Trace`
+is the `app.Controller`.
+
+Characteristics:
+
+All HTTP Methods are supported, for example if want to serve `GET`
+then the controller should have a function named `Get()`,
+you can define more than one method function to serve in the same Controller struct.
+
+Persistence data inside your Controller struct (share data between requests)
+via `iris:"persistence"` tag right to the field.
+
+Optional `Init` function to perform any initialization before the methods,
+useful to call middlewares or when many methods use the same collection of data.
+
+Access to the request path parameters via the `Params` field.
+
+Access to the template file that should be rendered via the `Tmpl` field.
+
+Access to the template data that should be rendered inside
+the template file via `Data` field.
+
+Access to the template layout via the `Layout` field.
+
+Access to the low-level `context.Context` via the `Ctx` field.
+
+Flow as you used to, `Controllers` can be registered to any `Party`,
+including Subdomains, the Party's begin and done handlers work as expected.
+
+Example Code:
+
+
+    // file: main.go
+
+    package main
+
+    import (
+        "github.com/kataras/iris"
+
+        "controllers"
+    )
+
+    func main() {
+        app := iris.New()
+        app.RegisterView(iris.HTML("./views", ".html"))
+
+        app.Controller("/", new(controllers.Index))
+
+        // http://localhost:8080/
+        app.Run(iris.Addr(":8080"))
+    }
+
+
+    // file: controllers/index.go
+
+    package controllers
+
+    import (
+        "github.com/kataras/iris/core/router"
+    )
+
+    // Index is our index example controller.
+    type Index struct {
+        router.Controller
+        // if you're using go1.9:
+        // you can omit the /core/router import statement
+        // and just use the `iris.Controller` instead.
+    }
+
+    // will handle GET method on http://localhost:8080/
+    func (c *Index) Get() {
+        c.Tmpl = "index.html"
+        c.Data["title"] = "Index page"
+        c.Data["message"] = "Hello world!"
+    }
+
+    // will handle POST method on http://localhost:8080/
+    func (c *Index) Post() {}
+
+
+Tip: declare a func(c *Index) All() {} or Any() to register all HTTP Methods.
+
+A full example can be found at: https://github.com/kataras/iris/tree/master/_examples/routing/mvc.
+
+
 Parameterized Path
 
 At the previous example,
@@ -1266,6 +1354,106 @@ Running the example:
 
     $ curl -s --cookie "mycookiesessionnameid=MTQ4NzE5Mz..." http://localhost:8080/secret
     The cake is a lie!
+
+
+Sessions persistence can be achieved using one (or more) `sessiondb`.
+
+Example Code:
+
+    package main
+
+    import (
+        "time"
+
+        "github.com/kataras/iris"
+        "github.com/kataras/iris/context"
+
+        "github.com/kataras/iris/sessions"
+        "github.com/kataras/iris/sessions/sessiondb/boltdb" // <- IMPORTANT
+    )
+
+    func main() {
+        db, _ := boltdb.New("./sessions/sessions.db", 0666, "users")
+        // use different go routines to sync the database
+        db.Async(true)
+
+        // close and unlock the database when control+C/cmd+C pressed
+        iris.RegisterOnInterrupt(func() {
+            db.Close()
+        })
+
+        sess := sessions.New(sessions.Config{
+            Cookie:  "sessionscookieid",
+            Expires: 45 * time.Minute, // <=0 means unlimited life
+        })
+
+        //
+        // IMPORTANT:
+        //
+        sess.UseDatabase(db)
+
+        // the rest of the code stays the same.
+        app := iris.New()
+
+        app.Get("/", func(ctx context.Context) {
+            ctx.Writef("You should navigate to the /set, /get, /delete, /clear,/destroy instead")
+        })
+        app.Get("/set", func(ctx context.Context) {
+            s := sess.Start(ctx)
+            //set session values
+            s.Set("name", "iris")
+
+            //test if setted here
+            ctx.Writef("All ok session setted to: %s", s.GetString("name"))
+        })
+
+        app.Get("/set/{key}/{value}", func(ctx context.Context) {
+            key, value := ctx.Params().Get("key"), ctx.Params().Get("value")
+            s := sess.Start(ctx)
+            // set session values
+            s.Set(key, value)
+
+            // test if setted here
+            ctx.Writef("All ok session setted to: %s", s.GetString(key))
+        })
+
+        app.Get("/get", func(ctx context.Context) {
+            // get a specific key, as string, if no found returns just an empty string
+            name := sess.Start(ctx).GetString("name")
+
+            ctx.Writef("The name on the /set was: %s", name)
+        })
+
+        app.Get("/get/{key}", func(ctx context.Context) {
+            // get a specific key, as string, if no found returns just an empty string
+            name := sess.Start(ctx).GetString(ctx.Params().Get("key"))
+
+            ctx.Writef("The name on the /set was: %s", name)
+        })
+
+        app.Get("/delete", func(ctx context.Context) {
+            // delete a specific key
+            sess.Start(ctx).Delete("name")
+        })
+
+        app.Get("/clear", func(ctx context.Context) {
+            // removes all entries
+            sess.Start(ctx).Clear()
+        })
+
+        app.Get("/destroy", func(ctx context.Context) {
+            //destroy, removes the entire session data and cookie
+            sess.Destroy(ctx)
+        })
+
+        app.Get("/update", func(ctx context.Context) {
+            // updates expire date with a new date
+            sess.ShiftExpiraton(ctx)
+        })
+
+        app.Run(iris.Addr(":8080"))
+    }
+
 
 More examples: