+++ sessions/sessiondb/redis/service/config.go
@@ -11,8 +11,6 @@ const (
 	DefaultRedisAddr = "127.0.0.1:6379"
 	// DefaultRedisIdleTimeout the redis idle timeout option, time.Duration(5) * time.Minute
 	DefaultRedisIdleTimeout = time.Duration(5) * time.Minute
-	// DefaultRedisMaxAgeSeconds the redis storage last parameter (SETEX), 31556926.0 (1 year)
-	DefaultRedisMaxAgeSeconds = 31556926.0 //1 year
 )
 
 // Config the redis configuration used inside sessions
@@ -33,21 +31,18 @@ type Config struct {
 	IdleTimeout time.Duration
 	// Prefix "myprefix-for-this-website". Default ""
 	Prefix string
-	// MaxAgeSeconds how much long the redis should keep the session in seconds. Default 31556926.0 (1 year)
-	MaxAgeSeconds int
 }
 
 // DefaultConfig returns the default configuration for Redis service.
 func DefaultConfig() Config {
 	return Config{
-		Network:       DefaultRedisNetwork,
-		Addr:          DefaultRedisAddr,
-		Password:      "",
-		Database:      "",
-		MaxIdle:       0,
-		MaxActive:     0,
-		IdleTimeout:   DefaultRedisIdleTimeout,
-		Prefix:        "",
-		MaxAgeSeconds: DefaultRedisMaxAgeSeconds,
+		Network:     DefaultRedisNetwork,
+		Addr:        DefaultRedisAddr,
+		Password:    "",
+		Database:    "",
+		MaxIdle:     0,
+		MaxActive:   0,
+		IdleTimeout: DefaultRedisIdleTimeout,
+		Prefix:      "",
 	}
 }